+++ www/inc/lib/pmd_root_page.php
@@ -140,6 +140,7 @@ class PMD_Root_Page extends PMD_Root{
 		$this->SetHeadJavascript("var pmd_url_static='{$this->conf['urls']['static']}';");
 		$this->SetHeadJavascript("var pmd_url_www='{$this->conf['urls']['www']}';");
 		$this->SetHeadJavascript("var pmd_lang='{$this->conf['app']['lang']}';");
+		$this->SetHeadJavascript("var pmd_url_home='{$this->conf['urls']['home']}';");
 
 	}