+++ www/inc/page/photos.php
@@ -31,7 +31,7 @@ class PMD_Page extends PMD_Root_Page{
 			$page['title']		=ucfirst($cur_album['title'])." - ".$this->lang['global']['menu_head']['photos'];
 						
 			if(isset($_GET['back'])){
-				$page['url_back']=$this->conf['urls']['www'];
+				$page['url_back']=$this->conf['urls']['home'];
 			}
 			else{
 				$page['url_back']="?selected={$data['selected']}#a_{$data['selected']}";