+++ www/inc/doc/changelog.md
@@ -7,6 +7,7 @@
 - new: Add 'Value' buttons, used by new 'therm' type
 - new: Selectors are shown as Popup Menus, when there is too much items to list (user-defined)
 - new: OpenWrt page: allows to Copy MAC address
+- new: Separate Dashboard page per client's host or IP address